+++ tests/unit/test_self_check_outdated.py
@@ -6,90 +6,16 @@ import sys
 import freezegun
 import pretend
 import pytest
-from mock import patch
 from pip._vendor import pkg_resources
 
 from pip._internal import self_outdated_check
 from pip._internal.models.candidate import InstallationCandidate
-from pip._internal.network.session import PipSession
 from pip._internal.self_outdated_check import (
     SelfCheckState,
     logger,
-    make_link_collector,
     pip_self_version_check,
 )
 from tests.lib.path import Path
-
-
-@pytest.mark.parametrize(
-    'find_links, no_index, suppress_no_index, expected', [
-        (['link1'], False, False,
-         (['link1'], ['default_url', 'url1', 'url2'])),
-        (['link1'], False, True, (['link1'], ['default_url', 'url1', 'url2'])),
-        (['link1'], True, False, (['link1'], [])),
-        # Passing suppress_no_index=True suppresses no_index=True.
-        (['link1'], True, True, (['link1'], ['default_url', 'url1', 'url2'])),
-        # Test options.find_links=False.
-        (False, False, False, ([], ['default_url', 'url1', 'url2'])),
-    ],
-)
-def test_make_link_collector(
-    find_links, no_index, suppress_no_index, expected,
-):
-    """
-    :param expected: the expected (find_links, index_urls) values.
-    """
-    expected_find_links, expected_index_urls = expected
-    session = PipSession()
-    options = pretend.stub(
-        find_links=find_links,
-        index_url='default_url',
-        extra_index_urls=['url1', 'url2'],
-        no_index=no_index,
-    )
-    link_collector = make_link_collector(
-        session, options=options, suppress_no_index=suppress_no_index,
-    )
-
-    assert link_collector.session is session
-
-    search_scope = link_collector.search_scope
-    assert search_scope.find_links == expected_find_links
-    assert search_scope.index_urls == expected_index_urls
-
-
-@patch('pip._internal.utils.misc.expanduser')
-def test_make_link_collector__find_links_expansion(mock_expanduser, tmpdir):
-    """
-    Test "~" expansion in --find-links paths.
-    """
-    # This is a mock version of expanduser() that expands "~" to the tmpdir.
-    def expand_path(path):
-        if path.startswith('~/'):
-            path = os.path.join(tmpdir, path[2:])
-        return path
-
-    mock_expanduser.side_effect = expand_path
-
-    session = PipSession()
-    options = pretend.stub(
-        find_links=['~/temp1', '~/temp2'],
-        index_url='default_url',
-        extra_index_urls=[],
-        no_index=False,
-    )
-    # Only create temp2 and not temp1 to test that "~" expansion only occurs
-    # when the directory exists.
-    temp2_dir = os.path.join(tmpdir, 'temp2')
-    os.mkdir(temp2_dir)
-
-    link_collector = make_link_collector(session, options=options)
-
-    search_scope = link_collector.search_scope
-    # Only ~/temp2 gets expanded. Also, the path is normalized when expanded.
-    expected_temp2_dir = os.path.normcase(temp2_dir)
-    assert search_scope.find_links == ['~/temp1', expected_temp2_dir]
-    assert search_scope.index_urls == ['default_url']
 
 
 class MockBestCandidateResult(object):
